pyproject.toml
[tool.poetry]
name = "trendy-orchestrator"
description = "Autonomous AI agent orchestrator for TrendyReports"
version = "0.1.0"
authors = ["Jerry"]

[tool.poetry.dependencies]
python = "^3.11"
anthropic = "^0.42.0"
psycopg = {extras = ["binary"], version = "^3.2"}
httpx = "^0.27"
pygithub = "^2.3"
schedule = "^1.2"
python-dotenv = "^1.0"
pydantic = "^2.0"
pydantic-settings = "^2.0"

[tool.poetry.scripts]
orchestrator = "orchestrator.main:main"
intake = "orchestrator.intake_server:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
